producer/requirements.txt
kafka-python
opencv-python
python-dotenv
